Fix import of callback url function and move function to utils libs
