+++ libs/designer-ui/src/lib/overview/__test__/utils.spec.ts
@@ -1,56 +1,7 @@
-import type { CallbackInfo, Run, RunError } from '../types';
-import { getCallbackUrl, isRunError, mapToRunItem } from '../utils';
+import { isRunError, mapToRunItem } from '../utils';
+import type { Run, RunError } from '@microsoft/utils-logic-apps';
 
 describe('lib/overview/utils', () => {
-  describe('getCallbackUrl', () => {
-    it('should return undefined when passed nothing', () => {
-      expect(getCallbackUrl(undefined)).toBeUndefined();
-    });
-
-    it('should return a callback URL when value is set', () => {
-      const callbackInfo: CallbackInfo = {
-        value: 'value',
-      };
-      expect(getCallbackUrl(callbackInfo)).toBe(callbackInfo.value);
-    });
-
-    it('should return a callback URL when relativePath is set', () => {
-      const callbackInfo: CallbackInfo = {
-        method: 'POST',
-        relativePath: '/apis/list',
-      };
-      expect(getCallbackUrl(callbackInfo)).toBe('/apis/list');
-    });
-
-    it('should return a callback URL when relativePath is set without prepended /', () => {
-      const callbackInfo: CallbackInfo = {
-        method: 'POST',
-        relativePath: 'apis/list',
-      };
-      expect(getCallbackUrl(callbackInfo)).toBe('/apis/list');
-    });
-
-    it('should return a callback URL when relativePath and basePath are set', () => {
-      const callbackInfo: CallbackInfo = {
-        basePath: '/root',
-        method: 'POST',
-        relativePath: '/apis/list',
-      };
-      expect(getCallbackUrl(callbackInfo)).toBe('/root/apis/list');
-    });
-
-    it('should return a callback URL when relativePath and queries are set', () => {
-      const callbackInfo: CallbackInfo = {
-        method: 'POST',
-        queries: {
-          'api-version': '2022-02-01',
-        },
-        relativePath: '/apis/list',
-      };
-      expect(getCallbackUrl(callbackInfo)).toBe('/apis/list?api-version=2022-02-01');
-    });
-  });
-
   describe('isRunError', () => {
     it('should return true when passed a run error object', () => {
       const runError: RunError = {
